Move actions to the default factory
