--- packages/core/src/actions/index.ts
@@ -1,6 +1,0 @@
-export * from './deploy'
-export * from './ekubo'
-export * from './launch'
-export * from './liquidity'
-export * from './memecoin'
-export * from './token'
