+++ packages/core/src/factory/default.test.ts
@@ -3,7 +3,7 @@ import { constants, RpcProvider } from 'starknet'
 import { describe, expect, test } from 'vitest'
 
 import * as TestData from '../../test/TestData'
-import { getPairPrice } from '../actions/token'
+import { getPairPrice } from '../utils/token'
 import { Factory } from './default'
 
 const provider = new RpcProvider({
