+++ packages/core/src/index.ts
@@ -1,5 +1,4 @@
-export * as actions from './actions'
 export * as constants from './constants'
 export * from './factory'
 export * from './types'
-export * from './utils/ekubo'
+export * from './utils'
